# extracts/nrf-openapi-extract.yaml
openapi: 3.0.0
info:
  version: '0.1'
  title: NRF Extracts

paths:
  /nnrf-nfm/v1/nf-instances/{nfInstanceID}:
    put:
      operationId: RegisterNFInstance
      parameters:
      - name: nfInstanceID
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/NfInstanceId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NFProfile'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NFProfile'
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NFProfile'
        '400':
          $ref: '#/components/responses/400'
                 

  /nnrf-disc/v1/nf-instances:
    get:
      summary: Search a collection of NF Instances
      operationId: SearchNFInstances
      parameters:
      - name: target-nf-type
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/NFType'
      - name: requester-nf-type
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/NFType'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
        '400':
          $ref: '#/components/responses/400'

components:
  schemas:
    SearchResult:
      description: Contains the list of NF Profiles returned in a Discovery response
      type: object
      required:
      - validityPeriod
      - nfInstances
      properties:
        validityPeriod:
          type: integer
        nfInstances:
          type: array
          items:
            $ref: '#/components/schemas/NFProfile'
    NFProfile:
      type: object
      required:
      - nfInstanceId
      - nfType
      - nfStatus
      - fqdn
      properties:
        nfInstanceId:
          $ref: '#/components/schemas/NfInstanceId'
        nfType:
          $ref: '#/components/schemas/NFType'
        nfStatus:
          $ref: '#/components/schemas/NFStatus'
        fqdn:
          $ref: '#/components/schemas/Fqdn'
    NFType:
      type: string
      enum:
        - UDM
        - AMF
        - AUSF
        - UDR
    ProblemDetails:
      type: object
      properties:
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
    NfInstanceId:
      type: string
      format: uuid
    Fqdn:
      type: string
      pattern: ^([0-9A-Za-z]([-0-9A-Za-z]{0,61}[0-9A-Za-z])?\.)+[A-Za-z]{2,63}\.?$
      minLength: 4
      maxLength: 253
    NFStatus:
      type: string
      enum:
        - REGISTERED
        - SUSPENDED
  responses:
    '400':
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'